AutoLM.jl Documentation

Data loading

AutoLM.load_imgs — Method

load_imgs(path, dims)

Reads all the volume images of an entire directory. Data must be stored as follows: The required input "path" is the path to a directory containing one sub-directory for every volume that has to be read. Each sub-directory contains the images making up one volume in .tif format. if the image names contain alphabetical characters (e.g. image001.tif, image002.tif ect.) the variable "numerical" has to be set to false, if the image names are just numerical values (e.g 1.tif, 2.tif, 3.tif, ect), it has to be set to true. Any file not ending in .tif will not be read.

Data will be stored in a 4D array [a x b c x n] with n being the number of sub-directories, a and b the resolution of the .tif images and c the number of images. The the resolution of the images has to be the same over all the volumes.

AutoLM.read_landmarks — Method

read_landmarks(path, num_landmarks, group)

Specific function to read avizo landmark data into an array. Reads all the files ending in .Ascii in the directory specified as "path". Data will be stored in a 2D array [c x n] with c being the number of individual 3D coordinates (30 coordinates in the case of 10 landmarks) and n being the number of landmark files read.

The variable Group specifies the group of landmarks in the Avizo file that has to be read (e.g "@1" in the case of group 1)

Data augmentation

AutoLM.flip_volume_front — Method

flip_volume_front(x, y)

Flip a volume so that the longitudinal (x) axis becomes the vertical (z) axis and adjust the landmakr coordinates accordingly.

AutoLM.flip_volume_side — Method

flip_volume_side(x, y)

Flip a volume so that the lateral (y) axis becomes the vertical (z) axis and adjust the landmark coordinates accordingly.

AutoLM.mirror_vol — Method

function mirror_vol(x, y)

A form of data augmentation. Mirrors a volume (exchanges x and y axis) and returns concatenation of original data and mirrored data.

Missing docstring.

Missing docstring for AutoLM.flip_2D(x, y). Check Documenter's build log for details.

AutoLM.flip_3D — Method

flip_3D(x, y)

Takes 3D volumes in a 4D tensor and landmark data in 2D tensor and returns the original volumes together with the flipped (clockwise) volumes, aswell as the coordinates for the original and the flipped volumes.

AutoLM.jitter_3D — Method

jitter_3D(x, y)

Jitters around binarized volumes based on their corresponding x/y landmarks, so that the relevant object (smaller voxel value) will still be fully inside the volume but moved around randomly inside the volume.

AutoLM.rotate_images — Method

rotate_images(img, lms, deg)

Takes a 3d tensor [res1 x res1 x n] with n images and rotates all of them in a counterclockwise direction around their center by deg degrees. Will adjust the 2d coordinates of the landmark array [coords x n]. Returns rotated images and adjusted landmarks.

AutoLM.rotate_volumes — Method

rotate_volumes(vols, lms, deg)

Takes a 4d tensor [res1 x res2 x res3 x n] with n volumes and rotates all of them around their center by deg degrees. The rotation is along the z-axis, so the z-coordinates will not be affected. x- and y coordinates of the landmark array [coords x n] will be adjusted. Returns rotated volumes and adjusted landmarks.